chore(PPDSC-2775): update accordion default styles (#602)
* fix(PPDSC-2775): update icon styles

* fix(PPDSC-2775): design review

* fix(PPDSC-2775): design reviews

* chore(PPDSC-2412): audio player styling changes (#592)

* chore(PPDSC-2412): audio player styling changes

* chore(PPDSC-2412): linting and comments

* chore(PPDSC-2412): typo

* chore(PPDSC-2412): typo

* chore(PPDSC-2412): simplify style presets

* chore(PPDSC-2412): revert typo

* chore(PPDSC-2412): update snapshots

* chore(PPDSC-2412): add codemod to apply all style changes

* chore(PPDSC-2412): update readme

* chore(PPDSC-2775): codemod for accordion

* fix(PPDSC-2775): revert readme

* fix(PPDSC-2775): lint and snapshots

* fix(PPDSC-2775): update codemods

* fix(PPDSC-2775): lint

* fix(PPDSC-2775): lint

* fix(PPDSC-2775): remove fontColorChange from button

---------
