+++ lib/codemod/src/v7.0.0/__tests__/all-default.test.js
@@ -1,13 +1,13 @@
 const jscodeshift = require('jscodeshift');
-const updateAllDefaultStyles = require('../all-default-styles');
+const updateAllDefault = require('../all-default');
 const dedent = require('dedent-js');
 
-describe('all-default-styles', () => {
+describe('all-default', () => {
   let source;
   let expected;
   let actual;
 
-  test('applies audio player default style updates', () => {
+  test('applies audio player default play pause button size', () => {
     source = dedent`
       const componentName = () => <AudioPlayerPlayPauseButton />;
     `;
@@ -16,7 +16,7 @@ describe('all-default-styles', () => {
       const componentName = () => <AudioPlayerPlayPauseButton size="large" />;
     `;
 
-    actual = updateAllDefaultStyles({source}, {jscodeshift});
+    actual = updateAllDefault({source}, {jscodeshift});
 
     expect(actual).toEqual(expected);
   });
