+++ lib/codemod/src/v7.0.0/all-default.js
@@ -1,6 +1,6 @@
-const updateAudioPlayerDefaultStyles = require('./audio-player-default-styles');
+const updateDefaultAudioPlayerPlayPauseButtonSize = require('./audio-player-play-pause-button');
 
-const codemods = [updateAudioPlayerDefaultStyles];
+const codemods = [updateDefaultAudioPlayerPlayPauseButtonSize];
 
 module.exports = function transformer(file, api) {
   let source = file.source;
